pyproject.toml
[tool.poetry]
name = "component-hub"
version = "0.1.0"
description = ""
authors = ["VSHN AG <info@vshn.ch>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
cookiecutter = "^1.7.3"
PyYAML = "^5.4.1"
PyGithub = "^1.55"
click = "^8.0.1"
importlib-metadata = "^4.5.0"
python-dotenv = "^0.20.0"
Jinja2 = "^3.0.1"
alt-profanity-check = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
mypy = "^0.941"
types-PyYAML = "^6.0.0"
autopep8 = "^1.5.7"
pylint = "^2.8.3"
flake8 = "^5.0.0"
black = "^21.6b0"
bandit = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
component-hub = 'component_hub.cli:main'

[tool.black]
line-length = 100

[flake8]
exclude = "*.egg*,.git,.tox,venv"
max-line-length = 120

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.MASTER]
output-format = "colorized"
reports = "no"
disable = "invalid-name, missing-function-docstring,missing-class-docstring,missing-module-docstring,fixme"
min-public-methods = 1

[tool.bandit]
exclude = ".cache,.git,.tox,build,dist,docs,tests"
targets = "."
